check size of cookie file
The documentation notes that the cookie file is always 32 bytes, and
requires that any other size file not be used as an auth cookie.

This is a very basic guard against tor asking for any file as a cookie.
It's not sufficient to really secure the cookie auth method. This is a
stopgap measure until SECURECOOKIE can be implemented.